gpg_tool/requirements.txt
python-gnupg==0.4.1